LayoutTests/fast/visual-viewport/visual-viewport-same-object.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gc.js"></script>
<script>
if (window.internals)
    internals.settings.setVisualViewportEnabled(true);

test(() => {
    window.visualViewport.newValue = 23;
    gc();
    assert_equals(window.visualViewport.newValue, 23);
}, "window.visualViewport is SameObject");
</script>
