LayoutTests/fast/visual-viewport/visual-viewport-same-object-expected.txt


PASS window.visualViewport is SameObject
